// src/libs/client/model/employeeDetailDto.ts
/**
 * SophieHR.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EmployeeAvatarDetail } from './employeeAvatarDetail';
import { EmployeeAddress } from './employeeAddress';
import { DepartmentIdNameDto } from './departmentIdNameDto';
import { CompanyIdNameDto } from './companyIdNameDto';


export interface EmployeeDetailDto { 
    id?: string;
    title?: string | null;
    gender?: string | null;
    userName?: string | null;
    firstName?: string | null;
    middleName?: string | null;
    lastName?: string | null;
    workEmailAddress?: string | null;
    personalEmailAddress?: string | null;
    workPhoneNumber?: string | null;
    workMobileNumber?: string | null;
    personalMobileNumber?: string | null;
    phoneNumber?: string | null;
    holidayAllowance?: number;
    jobTitle?: string | null;
    dateOfBirth?: string;
    startOfEmployment?: string;
    address?: EmployeeAddress;
    managerId?: string | null;
    avatar?: EmployeeAvatarDetail;
    department?: DepartmentIdNameDto;
    company?: CompanyIdNameDto;
    passportNumber?: string | null;
    nationalInsuranceNumber?: string | null;
}