// src/libs/client/model/employeeListDto.ts
/**
 * SophieHR.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EmployeeAddress } from './employeeAddress';
import { DepartmentIdNameDto } from './departmentIdNameDto';
import { CompanyIdNameDto } from './companyIdNameDto';


export interface EmployeeListDto { 
    id?: string;
    firstName?: string | null;
    middleName?: string | null;
    lastName?: string | null;
    jobTitle?: string | null;
    workEmailAddress?: string | null;
    personalEmailAddress?: string | null;
    workPhoneNumber?: string | null;
    workMobileNumber?: string | null;
    personalMobileNumber?: string | null;
    holidayAllowance?: number;
    dateOfBirth?: string;
    startOfEmployment?: string;
    address?: EmployeeAddress;
    departmentId?: string;
    companyId?: string;
    department?: DepartmentIdNameDto;
    company?: CompanyIdNameDto;
}